refactor: MapEditorService 책임 분리 (#78)
